# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pukaist Engine (Local-First)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/upload": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Upload File",
        "operationId": "upload_file_upload_post",
        "parameters": [
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional theme/tag for routing",
              "title": "Theme"
            },
            "description": "Optional theme/tag for routing"
          },
          {
            "name": "enqueue",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, enqueue the task",
              "default": true,
              "title": "Enqueue"
            },
            "description": "If true, enqueue the task"
          },
          {
            "name": "dedupe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, reuse existing uploads by hash instead of storing/enqueuing again",
              "default": true,
              "title": "Dedupe"
            },
            "description": "If true, reuse existing uploads by hash instead of storing/enqueuing again"
          },
          {
            "name": "intent",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional JSON intent (allowed_exts, prefilter_keywords, prefilter_min_chars, llm_mode, summary_enabled)",
              "title": "Intent"
            },
            "description": "Optional JSON intent (allowed_exts, prefilter_keywords, prefilter_min_chars, llm_mode, summary_enabled)"
          },
          {
            "name": "callback_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Webhook to receive job status",
              "title": "Callback Url"
            },
            "description": "Webhook to receive job status"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/docs": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Upload Doc Alias",
        "description": "Alias for /upload so the UI can POST /docs with multipart form-data.",
        "operationId": "upload_doc_alias_docs_post",
        "parameters": [
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional theme/tag for routing",
              "title": "Theme"
            },
            "description": "Optional theme/tag for routing"
          },
          {
            "name": "enqueue",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, enqueue the task",
              "default": true,
              "title": "Enqueue"
            },
            "description": "If true, enqueue the task"
          },
          {
            "name": "dedupe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, reuse existing uploads by hash instead of storing/enqueuing again",
              "default": true,
              "title": "Dedupe"
            },
            "description": "If true, reuse existing uploads by hash instead of storing/enqueuing again"
          },
          {
            "name": "intent",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional JSON intent (allowed_exts, prefilter_keywords, prefilter_min_chars, llm_mode, summary_enabled)",
              "title": "Intent"
            },
            "description": "Optional JSON intent (allowed_exts, prefilter_keywords, prefilter_min_chars, llm_mode, summary_enabled)"
          },
          {
            "name": "callback_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Webhook to receive job status",
              "title": "Callback Url"
            },
            "description": "Webhook to receive job status"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_doc_alias_docs_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "docs"
        ],
        "summary": "List Docs",
        "operationId": "list_docs_docs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional theme substring filter",
              "title": "Theme"
            },
            "description": "Optional theme substring filter"
          },
          {
            "name": "doc_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional doc_type substring filter",
              "title": "Doc Type"
            },
            "description": "Optional doc_type substring filter"
          },
          {
            "name": "label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional user_relevance filter: relevant|not_relevant",
              "title": "Label"
            },
            "description": "Optional user_relevance filter: relevant|not_relevant"
          },
          {
            "name": "review_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional review_status filter: reviewed|needs_follow_up",
              "title": "Review Status"
            },
            "description": "Optional review_status filter: reviewed|needs_follow_up"
          },
          {
            "name": "aoi_theme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional AOI theme filter (e.g., ALC_Confirmed, Modern_Treaty)",
              "title": "Aoi Theme"
            },
            "description": "Optional AOI theme filter (e.g., ALC_Confirmed, Modern_Treaty)"
          },
          {
            "name": "aoi_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional AOI code filter (ALCODE, TAG_ID, or SOI_ID)",
              "title": "Aoi Code"
            },
            "description": "Optional AOI code filter (ALCODE, TAG_ID, or SOI_ID)"
          },
          {
            "name": "aoi_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional AOI name substring filter",
              "title": "Aoi Name"
            },
            "description": "Optional AOI name substring filter"
          },
          {
            "name": "near_band_nbr",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional nearest First Nation office band number filter",
              "title": "Near Band Nbr"
            },
            "description": "Optional nearest First Nation office band number filter"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/signed-url": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Get Signed Url",
        "description": "Generate a pre-signed URL for client-side uploads.\nSupports S3/GCS/MinIO.",
        "operationId": "get_signed_url_signed_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/complete": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Complete Upload",
        "description": "Complete a client-side upload and trigger processing.",
        "operationId": "complete_upload_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_doc_alias_docs_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_doc_alias_docs_post"
      },
      "Body_upload_file_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_upload_post"
      },
      "DocsResponse": {
        "properties": {
          "docs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Docs"
          }
        },
        "type": "object",
        "required": [
          "docs"
        ],
        "title": "DocsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadResponse": {
        "properties": {
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "stored_as": {
            "type": "string",
            "title": "Stored As"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "enqueued": {
            "type": "boolean",
            "title": "Enqueued"
          },
          "deduped": {
            "type": "boolean",
            "title": "Deduped",
            "default": false
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "manifest_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "stored_as",
          "enqueued"
        ],
        "title": "UploadResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
...
  }
}